test/Makefile: explicitly compile object files from test source.
By passing the test t_*.c source file together with src/* object files in
the linkin step, no debugging symbols from CFLAGS gets added to t_*.c.

this is also not good if the compiler and linker is two different executables
(for example cc and ld, that gcc uses internally) as a linker would not understand C-code.

This patch makes sure that test sources is explicitly compiled, then linked.