Changing License to GPL v2.
The project directly includes code from the linux kernel, wich
are licensed under GPL version 2. however GPL Version 2 is not
compatible with Version 3 (there isn't much Version 3 is compatible
with anyway), therefore we are forced to use version 2.

The other reason is that we want users to have freedom to use code
from this project and since GPL Version 3 does not play nice
with almost any other license, that sure is a problem (I myself
actualy have this problem right now, with my own code).